refactor(vm): separate the warm key from the filesystem it fingerprints
The warm cache is a policy — which boot shapes are cacheable, what
identifies one, which keys to evict — wrapped around two stat calls. Only
the stat calls need a filesystem, so only they stay behind.

`policy/warm.rs` takes `WarmKey`, `FileFingerprint`, `warm_eligible`,
`reject_reserved_labels`, `WarmCache`, `MAX_WARM_KEYS`, `WARM_KEY_LABEL`
and their tests. `derive_warm_key` splits: `fingerprint(path)` reads the
inode in `warm.rs`, `warm_key(spec, kernel, rootfs)` derives the digest
purely — the same body, now callable without a rootfs on disk. `warm.rs`
keeps the catalog lookups and the publish, and re-exports the names its
callers already use.

No behavior change: the digest is byte-for-byte the one it was (same
fields, same order, same separators), and the three tests that need real
files or a catalog stay where they were. Two of them built a `WarmKey`
from a literal; they now derive one, since minting keys is the pure half's
job.
