+++ virt/arcbox-vm/src/sandbox/policy.rs
@@ -6,3 +6,4 @@
 //! decisions stay next to the resources they own.
 
 pub(super) mod pool;
+pub(super) mod warm;
